requirements.txt
sqlalchemy==2.0.31
pyjwt==2.8.0
pydantic==2.8.2
jose==1.0.0
pycryptodome==3.20.0
bcrypt==4.1.3
alembic==1.13.2
python-decouple==3.8
pytest==8.3.1
httpx==0.27.0
black==24.4.2
fastapi==0.111.1
fastapi-cli==0.0.4
google-generativeai==0.7.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31